requirements.lock.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
cssselect==1.3.0
distro==1.9.0
faiss-cpu==1.13.2
filelock==3.20.0
filetype==1.2.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.2.0
langchain-core==1.2.6
langchain-google-genai==4.1.2
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.14.0
networkx==3.6.1
numpy==1.26.4
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.1.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyMuPDF==1.26.7
pypdfium2==5.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
readability-lxml==0.8.4.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.2.2+cpu
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
watchdog==6.0.0
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
